Drop use of the CMake LOCATION target property
CMake 3.0 introduced policy CMP0026 to disallow use of the LOCATION
target property.  Drop our unused lookup of the protoc_compiler
LOCATION.  Replace our lookup of the kwProcessXML LOCATION with a direct
reference to the target in the custom command.  CMake will replace
the reference with the location while generating the project.

Change-Id: I78ffa82eb2aa57b30586e3f79102d2dfdaa07666