Drop outdated CMake Policy settings
Drop explicit calls to cmake_policy() to set policies that are already
set by our cmake_minimum_required(VERSION) calls.  Drop our setting of
CMP0017 to OLD because.  It was incorrectly set to OLD to silence a
warning but should have been set to NEW to silence the warning, and that
is now done by our cmake_minimum_required call.

Change-Id: I6cc32dd987510f34defc0d0b03e1925ce557ff4a