BUG #14673: Fix follow data.
Follow-data bug arose becase the "reset-camera" would happen before the
scene had updated the views, which is where correct time is passed to
the view.

In order to fix follow-data, we needed to add structure to when cues get
added. To make that possible, we merged vtkPVAnimationScene and
vtkSMAnimationScene since the two don't need to be separate anymore and
cleaned up the "Tick" logic to add this order specific update of cues.

Currently, I am doing that based on the type of the cue. Eventually, we
can add support for "priorities" or something like that it make it more
extensible.

Change-Id: If5c77b43b7b1b56d78e279d6a25cf0f0d39500a8