Fixing issue where view isn't visible that would crash ParaView.
We were dereferencing a null pointer when there was more than
a single layout when trying to make the thumbnail for the export
state wizard. This fixes that issue by checking on the pointer
value and if the view isn't visible it uses the label which
gives information on that.

Change-Id: I0bd929b5199b71e61669d6da044e9ce6e38a55d5