Simplify interfaces to add view frame buttons.
pqViewFrameActionGroup, pqViewFrameActionGroupInterface, and
pqViewFrameActionsBehavior were used in complicated way to provide a
very basic ability -- to support to adding buttons/actions to a
pqViewFrame. This commits refactors that logic.

We now have a new interace pqViewFrameActionsInterface (which no longer
requires a QActionGroup). This interface is used to add buttons to the
view-frame. pqStandardViewFrameActionsImplementation is the ParaView
specific implementation which manages decorations for all ParaView
views, renderviews, chartviews, spreadsheet views, etc.

Cleaning up pqSpreadSheetVisibilityBehavior to no longer manage the
decorations added to the view. Moved that
pqStandardViewFrameActionsImplementation as well. Now to customize the
view frames, applications only have to replace
pqStandardViewFrameActionsImplementation.

Change-Id: I1ecc1deefea010972257fc766d1485418d42c396