Fix saving/loading of settings files on server processes.
All server processes, including satellites in batch mode, were reading
and saving settings files. That was incorrect. The settings files need
to processed only on client processes. In symmetric batch mode too, we
read the file on the root node and then distribute it to everyone. This
commit fixes this logic.

Change-Id: I7835f3c1ccc7eb8369ec4cf14ba7d838ee7ee2b6