Attemping to fix windows EnSightGold reader issue in parallel.
The problem is stemming from  the fact that due to external data
changes, the sample data for EnSightGold (elements.geo) now has unix
style line endings on Windows. In the past, it had windows style line
endings. Opening the file as binary overcomes this issue. We don't need
to fix the readline calls since we seem to be using strncmp() everywhere
in this file. A better fix would be to update the entire EnSight reader
hierarchy to ensure that text files are opened as binary and "\r"
handled properly.

Change-Id: I6b2f04cb7cdd563b5ee047934080da3eb39eb517