web: update listFiles test for ExternalData
Now that the data is downloaded on-demand, the files expected of the
test may not exist if the GUI is not enabled (as it downloads the
majority of the data files). Instead, ensure the data exists here, but
select the files to download for being small to avoid downloading
megabytes of data purely for listing their names.

Change-Id: I80085ddb8ecb7cc0db12e10ea8496a41bb8c5699