Added documentation of setting up autoupdates for ParaViewWeb.
Documentation includes the full bash script which is now running on
pvw-test.kitware.com to keep the latest nightly binary version up
and running.

Also edited the quick start guide to point to the superbuild git
repo for downloading pvw-setup.py, so that the latest version will
always be retrieved.

Change-Id: Ib98c260e2f9b9129c8b1f3b3eb40dc845cbc80bd