Change how colors are stored in pqColorChooserButton.
Colors were stored in a QColor. Internally, QColor stores color
components as 16-bit unsigned ints. The Qt documentation of
QColor states the consequence of this:

"QColor supports floating point precision and provides floating point
versions of all the color components functions, e.g.getRgbF(), hueF()
and fromCmykF(). Note that since the components are stored using
16-bit integers, there might be minor deviations between the values
set using, for example, setRgbF() and the values returned by the
getRgbF() function due to rounding."

Because the value retrieved from the QColor might not be the same as
what was input, any color property passed through a QColor could be
slightly different from the default color property, causing the color
to be saved as a custom setting in the settings JSON file when doing
so wasn't warranted.

This is solved by storing the RGBA color components in
pqColorChooserButton as an array of doubles instead of in a QColor
object.

Change-Id: I750b991cbb17c3b8c4aeb6fb02b0c2936e0bca2f