Changed pvweb server side python to use new server manager api.
Removed LookupTableManager from helper and removed all uses from
protocols.  Added a new web protocol called ParaViewWebColorManager.
Added new tests to TestApp to test all of the new rpc methods in
the new protocol, also added documentation of all new methods to
the pvweb docs.

Made some changes to the structure of the javascript protocol
tests to suit a slight re-organization of the vtkweb testing code.
Basically, the asynchronous nature of the tests was causing
conflicts, so I had to ensure that all these tests run serially.

Changes to color management should be backwards compatible for
web applications, except that the new default color map is the
standard ParaView "cool" one, so a new baseline image was added
as well.

Change-Id: I19ea40b4c6e771225719853518f9d870adea6b8a