Fix incorrect setting of XML name for exposed properties
Exposed properties from a subproxy were *almost* able to support
default values different from the subproxy's definition, but not
quite. The XML name was overwritten by the exposed name, which caused
updates from the Qt side to not propagate to the server. This patch is
a little bit of a hack that resets the XML name after the exposed
proxy XML attributes have been processed.

Change-Id: I3a99a1104b16436be3882170c982ce3b8a341798