Use platform dependant separator for the PV_PLUGIN_PATH env. var
The problem is described in ticket #0014763.

The problem is that by some reason Paraview 4.1 only supports
semicolon (";") as a separator of the directories for the
PV_PLUGIN_PATH variable, though this is natural that separator
is platform-dependant (";" on Windows and ":" on Linux).
On the other hand ";" is used as a separator of the CMake list items.

Change-Id: I68767bce96fd6e8f8a4ae1cf33f01d62ccc7f336