Cleaned up JSON settings file when proxy group is empty
When a proxy group is empty, it is removed from the root JSON value.
Also, fixed a bug where plain old vtkSMProperties were not handled
correctly (they should not be written to/read from file). Did this by
adding boolean return values to SetPropertySetting() and
SetProxySetting() methods where true indicates the setting of values
was successful and false indicates failure.

Change-Id: I20bbd35a5dea3c63b4d62e2a31e7691c74564175