BUG 14787: Fixed error involving Python view
When ParaView is compiled with the PARAVIEW_ENABLE_MATPLOTLIB option
OFF, the XML for the Python view and representation are still
included while the class definitions for the view and representation
were not built or linked into the executable. This causes
problems in servermanager.createModule(), which iterates over proxy
XML definitions and attempts to create proxies defined in the XML.
Specifically, when servermanager.createModule() tries to create a
proxy for the python view, it fails because there is no instantiator
for the Python view.

Inclusion of the Python view and representation XML definitions
conditioned on PARAVIEW_ENABLE_MATPLOTLIB fixes this problem.

Change-Id: I29405016135c14f84c597e2abeab2cec565a548f