BUG #13804: Added support to show rendering information.
Added annotations to Render View to show information about the most
recent render including things like whether LOD was used, whether
remote rendering was used, and the frame rate.

To view this widget, go to the Settings dialog and on the "Render View"
tab, search of "annotation". Turn the "Set whether to see annotation..."
setting on to see this information popup in the render view.

Change-Id: I6986ee5964705470d061c18a2e875367cf3659af