Hiding non-reader sources for CoProcessing plugin in exporter.
In general, non-reader sources (e.g. Wavelet) won't have an input
association for the simulation but this can be confusing to new
users so by default I'm only going to show sources for the
export wizard that actually read from a file. There is an option
though to turn on all sources to make them available for the
situation where the source looks just like the simulation input.

Change-Id: I2ccac78c7f0ddd2ab45fcbe5eb0170dc6af4a0ee