Added buttons to save/restore defaults
Added buttons to save and restore default property values in the
properties, display, and view panels. Each panel has separate buttons
so that the settings for the proxy associated with each can be
saved/restored separately. This makes it possible to save settings for
a source without saving settings for the representation of that
source.

Change-Id: I53be2036de78905f59eb8524faa372bc4e38e65e