Introduce new classes for the new linked view Histogram representation.
This new view and the according representation allow to directly
plot the histogram of an input dataset without manually applying
the histogram filter on it. The representation manage the histogram
table creation and also manage selection: the user can select one
or more bars in the chart and it automatically selects the
elements in the other views (3D, spreadsheet, etc.) that falls
in the corresponding bins of the histogram. Note that the
reverse operation is not supported: selecting elements in another
view will not select corresponding histogram bars.

As the parallel coordinates, the histogram view supports only one
representation. It automatically switches representations to have
only one visible.

Change-Id: I481c9f8b27e8475175b503748db6686a143bc962