BUG 14830: Fixed initial display of opacity function editor
The initial range of the scalar opacity function was not set
correctly. Added a property "ScalarRangeInitialized" to the
PiecewiseFunction proxy that indicates whether the scalar opacity
function was initialized. This is the same property added to the
PVLookupTable proxy that makes the color lookup table initial display
work correctly. Its absence from the PiecewiseFunction proxy was
causing this bug.

Change-Id: I72096e947c871bee6a8438abdf207639202fffae