Add support for tracing view layout.
Added support for tracing view layout. Extended
vtkSMParaViewPipelineControllerWithRendering to automatically create
layout proxy if none exists. As a next step in this process, we should
update GUI to not automatically assign views to frames instead rely on
the servermanager to do it. That will make the code a lot simpler.

Change-Id: I3fd6554db83bfcda49200b62afbf5e690472ecbc