Fixed leaks in proxies due to tracing.
When tracing was active, the servermanager.Proxy instances would hang
around even after the proxy was delete resulting in the trace being
wrong. The issue came to forefront when the trace kept on recording
resetting of ranges for lut's that was used by deleted representations.
Fixed by ensuring that the accessors are cleaned up properly when
proxies are deleted.

Change-Id: Ib617f349691b37b891445817604cbc62a5c0ce7b