BUG: 13562. Fix default values selected for ArraySelectionDomain.
For readers with array selection domain, the default value of the
property was always empty i.e. the domain was never used to fill up the
default. As a result, when first "apply" happened, even if the user
didn't change the property manually, the property would be noted as
modified by the tracing logic since the setting of the default values
from the info property was happening on "apply". Fixed this by ensuring
that vtkSMArraySelectionDomain resets the associated property to its
default value which is the value reported by the info-property
associated with it.

Change-Id: Iec6028012b3de9444a1263f21e9f243f73de008f