Moved most of the code calling Python functions to Python side
The scripts that invoked the user-defined functions were getting long
and awkward to call from the C++ side, so they have moved to the
Python side.

In addition, added backwards compatibility support for calling
old-style render() functions so that no modifications to existing
scripts are required. This works by checking the number of arguments
required by the defined render() function. If the number of arguments
is two, an old-style render() function is assumed. If the number of
arguments is three, the new-style render() function is assumed.

Change-Id: I09d8a10d0fdc0573de6d6cfc7bcc8523c5fb43f1