BUG: #14863 - Fix path initialization when lib dir in parent of .app bundle
This fixes a problem where a "lib" directory was used to determine if
the .app directory is in a build directory.  Use ParaViewConfig.cmake
as a better heuristic (although the same problem would still exist if
that file existed in unexpected places).

Also refactor the if/else statements to be more clear and indicate an error
in the fall-through case that existed previously.

Change-Id: I843ed13409dfc022b2859ffbc41541dde281b1d9