Reset image data in the vtkPythonView before each render
If this is not done, the previous successfully generated image may
still be displayed.

In addition, the Python environment is now reset prior to evaluating
the python script to clear out any definitions set in prior
evaluations of different verions of the script. For example, one could
define a render() method in the script, then set the script to the
empty string. Without resetting the Python environment, the render()
method would still be defined, which is unexpected stateful behavior.

Change-Id: I8979d8c2f9ec1657c7a84292e18d77147d947647