Handle the case with mismatched arrays across ranks.
It's highly possible (e.g in Exodus with can.ex2) for all ranks to not
have the same set of arrays. This raises errors in evaluating expression
on certain ranks. Adding NoneArray for those missing arrays addresses
this issue.

Also fixed a warning in vtkMultiProcessControllerHelper exposed by
FindData test in parallel.

Change-Id: I7b137dbd8c52598ae33aab96e0895a2025ceb7c1