BUG #14874. Bring back "Scalar Opacity Unit Distance".
The "Scale" property shown on the old color map editor was really
setting the scalar opacity unit distance on the representation. Ensuring
that the "Scalar Opacity Unit Distance" property is exposed as an
advanced property on the "Properties" panel. Note, it's on the
Properties panel and not on the "Color Map Editor" panel.

Change-Id: I6cada928f9b21b3460679f7ba1de5911793cf42a