BUG #14834: Fix representation update logic.
vtkSMRepresentationProxy tried to take extra precations
when calling vtkSMProxy::PostUpdateData() to ensure that the method
wasn't called unless the VTK-side representation truly updated. If the
VTKRepresentation didn't update, the upstream pipeline may not be
updated and hence telling the input proxies that they have been updated
would be incorrect.

However it missed a case when loading state files. When loading state
files, the this->MarkedModified is false by default. Thus the first time
PostUpdateData() was called, it always assumed the VTKRepresentation had
updated since we used the same flag. Fixed that by adding a new flag
VTKRepresentationUpdated to indicate that the VTK representation has
updated.

This address the bug.

Change-Id: I8a74d1734e0cec6a0f17690551e9a01761791fa4