BUG #14771: Fix vtkAnnotateGlobalDataFilter.
Fix (and upgrade) vtkAnnotateGlobalDataFilter after numpy support
changes in VTK/ParaView. Also fixes the bug reported in BUG #14771.
Instead of putting the logic to come up with the expression in C++, we
add a new method in annotation.py. It's easier to do that determination
in Python. Restructured code from vtkPythonAnnotationFilter to make that
possible.

Change-Id: I027ee8321897b93a8e24401e7e5b190a6b55981d