Removed bool return value from LoadSettings()
The original intent was to indicate success or failure of loading
settings. However, it's not clear what success or failure means
because it is legitimate to not have a user- or site-settings file, so
indicating failure if one of these cannot be loaded is not really
accurate. Moreover, the return value wasn't used by the caller.

In addition, there was a bug in the previous implementation. If the
user settings weren't loaded, the site settings wouldn't be loaded
either. Now this method no longer returns any boolean value.

Change-Id: I039fc39aec3866e0d5b349fa43fc8f5c0d648d2a