Try to load site settings files and use return value in search logic
Now instead of checking whether a settings file exists before
attempting to load it, we just try to load it and use the return code
from vtkSMSettings::AddCollectionFromString() to determine whether to
search other locations. This required fixing some of the behavior of
vtkSMSetting::AddCollectionFromFile() when the given file couldn't be
loaded.

Change-Id: Ic4f1caea33454453a4c7049c10642148a91f9cad