BUG #14889. Fixed pqDisplayColorWidget for values not in domain.
Fixed pqDisplayColorWidget to properly handle the case when the values
for the SMProperty are not in the domain. This change also ensures
that whenever the domain is updated with relevant information, the UI
will also update to show most up-to-date information  about component
available, and the selected component.

Also added a test (LoadPartiallyUpdatedStateFiles) which tests both,
BUG #14888 and BUG #14889.

Change-Id: Id0488f0e024b7580aea5ef687f4efcd486ced400