BUG #14789: Added ability to save images with border betn views.
There're new settings under "Settings | General" in "Screenshot Options"
group that allow users to pick a border and color for the border to use
when saving screenshots. (By default, the border width it set to 0
pixels).

Updated LoadStateMultiView test to also test this new border
functionality.

The implementation extends vtkSMUtilities::Merge() to add a border as
images are merged. The border is added as an overlay rather than my
offsetting images so that we don't need to jagged edges or having to
resize views.

Change-Id: I1a7baa15e5756baec918629b875d6e26bd027168