Refactoring vtkPVGlyphFilter for better masking of points.
Refactoring vtkPVGlyphFilter to improve usability esp when glyphing
random points. The vtkPVGlyphFilter now uses a random spatial
distribution of points to decide which points in the dataset to glyph,
if requested.

The filter has 3 modes: glyph all points, glyph every n-th point, and
glyph a spatial distribution of points.

This commit also includes Qt side enhancemnets needed for this filter to
work. These include:

Added pqGenericPropertyWidgetDecorator to handle show/hide
enable/disable widgets when the value of some other property matches a
criteria. Glyph uses it to hide the properties that depend on the "Glyph
Mode".

Change-Id: Iadb659c380d9c58766f9d0996f0ffd4a018b16ba