Added ability for pqPropertyWidget to update when in advanced mode.
We have cases where a pqPropertyWidget need to change itself based on
whether the container pqProxyWidget is showing "advanced" properties or
not. For example, thq pqProxySelectionWidget which is used for
properties with proxy-list-domains, we may want to not show the
properties for the selected proxy in the domain, unless in advanced
view.

To facilitate that pqPropertyWidget now has a new method
"updateWidget()" which is called by pqProxyWidget before showing the
widget. Supclasses (such as pqProxyPropertyWidget) can override this
method to update their appearance based on the panel mode.

pqProxyPropertyWidget (working with pqProxySelectionWidget) overrides
this method to hide/show the widget for the selected proxy.

Change-Id: Ifdbbdfec1828b4eba717d79a9412c206a994bbd8