Added ability to easily set scale factor for glyph.
Added a button on the "Scale Factor" property widget to easily reset the
scale factor using current data value. This replaces the "automatic"
setting of scale factor that was done in pqGlyphPanel (which was a hack)
by a more acceptable solution for now the Properties panel works.

The "Reset Scale Factor" button also gets highlighted when the scale
factor domain value is potentially changed to make the user aware that
he may want to look at the scale factor.

Change-Id: I9d60816393c69f7c570ec85becf8c5559c7976af