vtkSMViewProxy: support transparent screenshots
The idea is basically to set the background color to black, take a
screenshot, then white and take another screenshot. From the difference
between the two images, we can determine whether the background is
showing through or not.

For translucency, the HSV value is calculated from each screenshot and
the alpha is the difference in the V component. The color values from
the black image are used since it doesn't change the input values at
all.

Change-Id: Icace035561660d6442688aeded622cac261d3886