Filter menu items are now enabled/disabled lazily
The enabled/disabled state of the Filters menu items used to be
calculated every time the selection in ParaView chagned.  This commit
causes these changes to instead set a dirty bit in the
pqFiltersMenuReaction and defers the actual computation of which menu
items are enabled or disabled until the menu is shown.  The menu's
aboutToShow signal and the new aboutToShowQuickLaunch signal on
pqPVApplicationCore are the only triggers to recompute the
enabled/disabled state of filter menu items.  QActions that are shown
in the toolbars are still updated at every signal since these are
already shown.

Change-Id: I08d60fc9008f1e593aa466bce6e8d243fee36f55