Add python syntax highlighting
This commit adds python syntax highlighting to the pqPythonScriptEditor
and the QTextEdit used to put in code for the Python Programmable
filters.  The pqPythonSyntaxHighlighter class is added and can be
attached to arbitrary QTextEdits to give python syntax highlighting and
some useful code editing properties like automatically expanding tabs
and a fixed-width font.

Change-Id: I1b558cb434f4ffd417a178a3c804019719df19a7