BUG #14915: Added ability to reset scalar value.
Added mechanism for the user to reset the scalar value based on the
domain explicitly, when the array selection is changed. We still don't
reset the value automatically, it's the user responsibility. But we mark
the as highlighted, making it easier for the user to realize that he can
reset the scalar value.

Change-Id: I6f42f542fb3914b79d550dce122eeee1a7a8f31f