BUG #14918. Fix color palette restore after save snapshot.
The vtkSMGlobalPropertiesProxy used for color palette's "SaveXMLState"
doesn't save the property values, but saves the linking state for global
properties. Hence, we were using it incorrectly to restore the property
values. Fixed that by creating a temporary copy of the color palette
before changing it and then using that to restore the property values.

Change-Id: Ia49e04e1c15fee477b34ebd5e9843259705c5cac