BUG #14913: Fix enabled state for 3D widgets.
vtk3DWidgetRepresentation and subclasses require that Enabled and
Visibility properties are explicitly managed by the client. Thus it was
possible to have cases where a widget is enabled, but not visible!
That's bad. This change ensures that a widget cannot be enabled, unless
its representation is visible.

Also simiplified logic in vtk3DWidgetRepresentation when dealing with
custom transforms.

Change-Id: I45373e1bfdffbd7cd4e0f5130290d91280fdc197