Add a dialog for properties panel python scripts
This commit adds a widget that pops its contents out into a dialog.
This is used for the properties panel text fields that contain python
scripts for the Programmable Filter and Programmable Source.  A button
is added to pop the text area out into the dialog and bring it back.

Change-Id: I1cd39479859046e35f7cf5af41a017b47461213b