Make Histogram View work in client-server mode.
vtkPVHistogramChartRepresentation did array name determination in a
routine that would execute only on the server-side (where data was
available). While the array name was needed on the client side to
create selections and for labeling. Fixed that.

Also avoid doing any work that modified MTime in methods that gets
called in RequestData(). That was not needed anyways. Removed that. This
avoid the bug where the selection shown would be cleared if the
histogram color was changed.

Change-Id: I8b3e09aad623dc2eae5072984795edba1bbc95e8