Added generic mechanism for views to show only 1 repr at a time.
If a view has a hint "ShowOneRepresentationAtATime", then it implies
that the view can only show 1 data-representation at a time.
vtkSMParaViewPipelineControllerWithRendering::Show() ensures that the
such views hide other representations that may be currently visible.
This is a better approach to handling this in Qt subclasses since now it
works as expected from Python as well as the undo/redo stack doesn't get
confused.

Change-Id: I33fd32c2d1e517ad16ff06f1526b852e4f89f12e