Refactoring transformation of selection for chart views.
All chart views/representations (not just the histogram view) need to
transform the vtkSelection created by the vtkChart to a type of
selection that qualifies the data going into the chart representations,
not simply that going into the view. For example, if we were plotting
cell data in a line chart view, all selections we were making were
essentially incorrect, since ParaView was thinking the view selecting
point-ids.

This commit fixes that by adding logic to transform selection made in
the view by collaborating with the vtkChartRepresentation before passing
that on to the UI.

This change leaves the current limitation when selecting multiple
representations in the same view unchanged. We will fix that at a later
stage (that, and also selections with composite datasets).

Change-Id: I38ec26416ba7f8c76678f320c24234b049b20410