Added ability for selection representations to use original selection.
There are cases (e.g. Histogram View/Representation) where the
selection-representation needs the original input selection (and not the
selection converted to IDs). Added mechanism for that.

vtkPVExtractSelection now has a 3rd output which is simply the same as
its inputs. vtkSMChartRepresentationProxy can pass either the 2nd or 3rd
output from vtkPVExtractSelection to the selection representation proxy
based on how the property was named.

Change-Id: I06ebeef9e9ff8168dacafe520e6c7d1202f2de59