Added support for two selection transformation.
When it comes to chart representations and selections, we need
bi-directional transformation from view to input data and vice-versa.
When a selection is created in the view, the application should get the
selection that's defined on the elements in the input dataset (not the
elements shown in the view). This is done using MapSelectionToInput()
(formerly called TransformSelection) on vtkChartRepresentation.

Then when a selection comes down the input pipelinen to
vtkChartSelectionRepresentation, it needs to convert it to map to the
elements in the view. This is done using the new
vtkChartRepresentation::MapSelectionToView().
vtkChartSelectionRepresentation calls this in RequestData() before
delivering the transformed selection to the rendering processes.

With this changes, BUG #0014947 is addressed.

Change-Id: Ia9a52cc011725b35a54a2ee695b8a3651fa31f67