BUG #14279. Fix random glyphs on Windows.
BUG #14279 was encountered because the code that used rand() didn't not
take into consideration that the int value returned by rand() may be
less than the number of point ids in the dataset. On Windows, is was
indeed limited to 32767 (on unixes, it wasn't that low). Thus resulted
in the rand point ids selected being limited to that range.
Fixed that.

Change-Id: Ifd0e13cdba8404776a2388e2c5c3f68c6889a8ca