Modify representation to handle same sized blocks
This commit modifies vtkStreamingParticlesRepresentation to handle
blocks of different resolutions being the same size by adding an 'amount
of detail' information type to the block information object and using
this to approximate how much detail will be visible.  This behavior is
now an option on the vtkStreamingParticlesRepresentation, although it
will not be used unless the source provides the new information key on
its data.

Change-Id: I9efe899618c7b17df4c09b2d64e2c6cb3f8bfe55