Fixing LoadDistributedPlugin() works for Catalyst.
This works for both the build directory and instal directory
for a PV development install, as long as the build directory
still exists. This is because it looks for the plugin
library in the build directory in both cases.
Also getting rid of trailing white space and a compile warning.

Change-Id: I1a3457f90f51eaf112bccba8fb949bcf4f7f3be7