Enhance geometry representation to support non convex polygon rendering.
A new "Triangulate" option is added to the representation.
The work is then done in the geometry filter. If the option is enabled
a vtkTriangleFilter is applied to the mesh (for unstructured & polydata
inputs). The support for picking and correct edges displayed is performed
using the same technique as with non linear cells - original cell ids
are used to compute boundary edges which are then associate to the
mesh point data as an EDGEFLAG field, used downstream by the mapper.

See mantis entry for test data:
http://www.paraview.org/Bug/view.php?id=14308

Tested in built-in, client, server and parallel server modes.

Change-Id: Ib9ab09f84b78a827ed94cbd92f667b49ac3beebb