BUG #14994. Fix issue with text source/representation in parallel.
With changes to the pipeline, vtkPVTextSource was only getting called on
the root-node. Hence text was not available on satellites.
vtkTextSourceRepresentation assumed that text was available on all nodes
to avoid redistribution of data. Hence we failed with missing text
errors when rendering on satellites. Fixed that by making sure that
vtkPVTextSource gets executed on all ranks.

Also fixed vtkTextSourceRepresentation to not raise errors when empty
text is set as text.

Update UserTransformOnRepresentation test to test text sources in batch
mode.

Change-Id: I049a6b94df8e3010d3eda8c31afa6b433362e7ad