Fix chart representation being modified in render.
vtkSMContextViewProxy was updating axes ranges at end of render which
resulted in making the representation proxy modified. While this is not
the end of the world in most cases, in symmetric batch mode, it can use
the view proxy to be marked 'dirty' differently on ranks (since axes
ranges never change on satellites.

This hack, similar to what we do to handle updates for camera properties
addresses that issue. We need a more gracefully support for such
'auto-updating' properties.

Added a test would fail with "pvbatch -sym" without this fix.

Change-Id: I18ae6c1efd5940855097f1ad8deeae527a0b846d