Fixed issues with plugins and static pvserver.
The problem setup is this:

I have two local builds of ParaView, one static and one shared.
First, I connect with shared pvserver as (cs://localhost) and load some
distributed plugin (e.g. Moments) -- works as expected.

Next, I restart ParaView and connect to the static pvserver using the
same url (cs://localhost). Now attempting to load the same plugin will fail.

This happens because the client tells the server to load the plugin
configuration XML from its QSettings (since we want to restore auto-load
preferences, as well as extra use-loaded plugins). The second time, it
uses the XML obtained from the pvserver ran the first time (which was a
shared build) and hence has different paths for the plugins.

This interim fix addresses this issue by not overriding the paths the
server had determined. We need a cleaner fix for plugin configuration
management.

connect with the shared pvserver

Change-Id: Iaa7ce40928e34418ac59df9d23b04e704375db5d