Add workaround to skip display test.
In some setups, the X server may not be ready for clients when the
pvserver starts, but would be by the time a render view is created.
Added a workaround for such cases. By specifying the
"--disable-xdisplay-test" command line option on the rendering server
processes, you can skip this test entirely.

Change-Id: I061618b755907d7633f1d57e222a4790f37cdafa