Fixing how Dolfin finds ParaView.
The previous version assumed that Dolfin was working with an
installed version of ParaView that had development files installed.
These changes allow Dolfin to work with either an installed location
of ParaView or the build tree location of ParaView. Also fixed the
velocity specification in the adaptor. VTK's format for vectors
is opposite of Dolfins.