BUG #15025: Fix toggling of text annotations.
Text annotations (and other widgets) didn't work property in parallel
batch mode. The problem resulted from the fact that the
"CurrentRenderer" on the 3D widget was getting unset by logic in
vtkAbstractWidget. A good way of avoiding that is to set the
DefaultRenderer on the widget-representation. Updated
vtk3DWidgetRepresentation accordingly.

Also added a test (AnnotationVisibility).

Change-Id: I0786cceed56ad9de682638ef675511cf088fb77b