Blocks sent from server to client are now deleted when not needed
This commit fixes a bug in vtkStreamingParticlesRepresentation that
prevented blocks from being purged from the dataset when no longer
needed if the client was doing the rendering and the server had the
data.  The deletion only happened on the server side.  The fix was to
add a field data to the dataset sent to the client to request purging of
blocks and change the vtkComposite(Reader|Writer) to support saving
field data.

Change-Id: I3d7cc0ad6978d7e25a3b8abc759a5e1203eff27a